+++ player/s98player.cpp
@@ -616,7 +616,7 @@ void S98Player::RefreshMuting(S98_CHIPDEV& chipDev, const PLR_MUTE_OPTS& muteOpt
 UINT8 S98Player::SetDeviceOptions(UINT32 id, const PLR_DEV_OPTS& devOpts)
 {
 	size_t optID = DeviceID2OptionID(id);
-	if (optID == (UINT32)-1)
+	if (optID == (size_t)-1)
 		return 0x80;	// bad device ID
 	
 	_devOpts[optID] = devOpts;
@@ -630,7 +630,7 @@ UINT8 S98Player::SetDeviceOptions(UINT32 id, const PLR_DEV_OPTS& devOpts)
 UINT8 S98Player::GetDeviceOptions(UINT32 id, PLR_DEV_OPTS& devOpts) const
 {
 	size_t optID = DeviceID2OptionID(id);
-	if (optID == (UINT32)-1)
+	if (optID == (size_t)-1)
 		return 0x80;	// bad device ID
 	
 	devOpts = _devOpts[optID];
@@ -640,7 +640,7 @@ UINT8 S98Player::GetDeviceOptions(UINT32 id, PLR_DEV_OPTS& devOpts) const
 UINT8 S98Player::SetDeviceMuting(UINT32 id, const PLR_MUTE_OPTS& muteOpts)
 {
 	size_t optID = DeviceID2OptionID(id);
-	if (optID == (UINT32)-1)
+	if (optID == (size_t)-1)
 		return 0x80;	// bad device ID
 	
 	_devOpts[optID].muteOpts = muteOpts;
@@ -654,7 +654,7 @@ UINT8 S98Player::SetDeviceMuting(UINT32 id, const PLR_MUTE_OPTS& muteOpts)
 UINT8 S98Player::GetDeviceMuting(UINT32 id, PLR_MUTE_OPTS& muteOpts) const
 {
 	size_t optID = DeviceID2OptionID(id);
-	if (optID == (UINT32)-1)
+	if (optID == (size_t)-1)
 		return 0x80;	// bad device ID
 	
 	muteOpts = _devOpts[optID].muteOpts;
